mandelbrot/requirements.txt
numpy
numba
PySide6
pytest
absl-py
